engine: Clean abstractions for RocksDB benchmarks
In anticipation of adding benchmarks for alternative Engine implementations.
